+++ packages/terra-application-layout/tests/wdio/routing-menu-spec.js
@@ -28,7 +28,7 @@ Terra.describeViewports('RoutingMenu', ['small'], () => {
     });
 
     it('successfully navigated away', () => {
-      expect(browser.url().value).to.equal('https://engineering.cerner.com/terra-ui/#/home/terra-ui/index');
+      expect(browser.url().value).to.equal('https://engineering.cerner.com/terra-ui/home/terra-ui/index');
     });
   });
 });