+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Updated to interface with react-intl's `FormattedMessage` rather than old React Context API.
+* updated package.json test scripts
 
 3.8.0 - (July 30, 2019)
 ------------------