+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -3,6 +3,23 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* updated package.json test scripts
+
+3.8.0 - (July 30, 2019)
+------------------
+### Removed
+* Removed DEPENDENCIES.md file
+
+3.7.1 - (July 23, 2019)
+------------------
+### Changed
+* Updated DEPENDENCIES.md doc
+
+3.7.0 - (July 23, 2019)
+------------------
+### Changed
+* Minor dependency version bump
 
 3.6.0 - (July 16, 2019)
 ------------------